backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
anthropic>=0.40.0
httpx
twilio
sendgrid
firebase-admin
sqlalchemy[asyncio]
asyncpg
alembic
redis[asyncio]
cryptography
numpy
pandas
scikit-learn
joblib
pytest
pytest-asyncio>=0.23
pytest-mock
pytest-benchmark
fakeredis[aioredis]
aiosqlite
websockets>=12.0
psycopg2-binary
